Changed configuration to point to local installables.
Former-commit-id: 085f07f7710252e8f619fbd4c4e79f31849dcf7d [formerly 0b556b541c739847bbfea8feccd1702d6f9aea1a] [formerly e99229da2259afb3aad028cbd1d37bcc4436a2e7 [formerly e99229da2259afb3aad028cbd1d37bcc4436a2e7 [formerly e99229da2259afb3aad028cbd1d37bcc4436a2e7 [formerly e99229da2259afb3aad028cbd1d37bcc4436a2e7 [formerly e99229da2259afb3aad028cbd1d37bcc4436a2e7 [formerly 3447b24c16a1b7b0bb1ec545edc68e35c741a3cf]]]]]]
Former-commit-id: e99229da2259afb3aad028cbd1d37bcc4436a2e7
Former-commit-id: 9b169e28528dac3d8540f2e2208c6d539b9ee337
Former-commit-id: 31e7f4235df965a5846cbd58cefe31cbdb3fcbe6 [formerly dfbb53e65a379a5c5c47c6fa922abf415aa7e5fe]
Former-commit-id: bf3291c27f3472f9c3e8b82bfb1a3772c5f0bb60
Former-commit-id: 1e58b1aa461223c4791d19dcd05283796d0998a5 [formerly ae44d9608dd6c6cb1a3d1e284ebe59bfeb8fedfa]
Former-commit-id: 00337b275d07c8f3b4fbdea8cbc811f8767113b3

+++ scripts/jenkins-setup/setup-centos.sh
@@ -2,24 +2,59 @@
 
 
 cd /opt;
-wget --no-cookies --no-check-certificate --header "Cookie: gpw_e24=http%3A%2F%2Fwww.oracle.com%2F; oraclelicense=accept-securebackup-cookie" "http://download.oracle.com/otn-pub/java/jdk/8u45-b14/jdk-8u45-linux-x64.rpm";
-#wget --no-cookies --no-check-certificate --header "Cookie: gpw_e24=http%3A%2F%2Fwww.oracle.com%2F; oraclelicense=accept-securebackup-cookie" "http://download.oracle.com/otn-pub/java/jdk/8u45-b14/jre-8u45-linux-x64.rpm";
+echo ...........................Installing java...........................
+if [ -f "/vshare/base-images/jdk/jdk-8u45-linux-x64.rpm" ]; then
+	yum install -y /vshare/base-images/jdk/jdk-8u45-linux-x64.rpm
+else
+	wget --no-cookies --no-check-certificate --header "Cookie: gpw_e24=http%3A%2F%2Fwww.oracle.com%2F; oraclelicense=accept-securebackup-cookie" "http://download.oracle.com/otn-pub/java/jdk/8u45-b14/jdk-8u45-linux-x64.rpm";
+	rpm -Uvh jdk-8u45-linux-x64.rpm;
+fi
 
-rpm -Uvh jdk-8u45-linux-x64.rpm;
-#rpm -Uvh jre-8u45-linux-x64.rpm;
+echo ...........................Installing sshpass...........................
+if [ -f "/vshare/base-images/misc/sshpass-1.05-9.2.x86_64.rpm" ]; then
+	yum --enablerepo=epel -y install /vshare/base-images/misc/sshpass-1.05-9.2.x86_64.rpm
+else
+	yum --enablerepo=epel -y install sshpass
+fi
 
-yum --enablerepo=epel -y install sshpass
+echo ...........................Installing git...........................
+if [ -f "/vshare/base-images/jenkins/git-1.8.3.1-5.el7.x86_64.rpm" ]; then
+	yum install -y git
 
-yum install -y git
+	#yum install -y /vshare/base-images/jenkins/perl-Error-0.17020-2.el7.noarch.rpm
+	#yum install -y /vshare/base-images/jenkins/perl-TermReadKey-2.30-20.el7.x86_64.rpm
+	#yum install -y /vshare/base-images/jenkins/libgnome-keyring-3.8.0-3.el7.x86_64.rpm
+	#yum install -y /vshare/base-images/jenkins/perl-Git-1.8.3.1-5.el7.noarch.rpm
+	#yum install -y /vshare/base-images/jenkins/git-1.8.3.1-5.el7.x86_64.rpm
+else
+	yum install -y git
+fi
 
-wget -O /etc/yum.repos.d/jenkins.repo http://pkg.jenkins-ci.org/redhat-stable/jenkins.repo
-rpm --import https://jenkins-ci.org/redhat/jenkins-ci.org.key
-yum -y install jenkins
+echo ...........................Installing jenkins...........................
+if [ -f "/vshare/base-images/jenkins/jenkins-ci.org.key" ]; then
+	rpm --import /vshare/base-images/jenkins/jenkins-ci.org.key
+else
+	wget -O /etc/yum.repos.d/jenkins.repo http://pkg.jenkins-ci.org/redhat-stable/jenkins.repo
+	rpm --import https://jenkins-ci.org/redhat/jenkins-ci.org.key
+fi
+
+if [ -f "/vshare/base-images/jenkins/jenkins-1.642.1-1.1.noarch.rpm" ]; then
+	yum -y install /vshare/base-images/jenkins/jenkins-1.642.1-1.1.noarch.rpm
+else
+	yum -y install jenkins
+fi
+
+
+echo ...........................Starting Jenkins...........................
 service jenkins start
-sleep 1m
+sleep 20
+
+echo ...........................Copying configuration files...........................
 cp /mnt/gluster/repo/jenkins-setup/plugins/*.hpi /var/lib/jenkins/plugins/
 chown jenkins:jenkins /var/lib/jenkins/plugins/*
 cp /mnt/gluster/repo/jenkins-setup/config/*.xml /var/lib/jenkins/
 chown jenkins:jenkins /var/lib/jenkins/*
 cp /mnt/gluster/repo/jenkins-setup/job-config/pwd /var/lib/jenkins/
+
+echo ...........................Restarting Jenkins...........................
 service jenkins restart
